:construction: Add toasts for user feedback when joining a quest

+++ frontend/src/graphql/joinQuestByCode/mutationJoinQuestByCode.ts
@@ -1,7 +1,7 @@
 import { gql } from "@apollo/client";
 
 export const mutationJoinQuestByCode = gql`
-  mutation JoinQuestByCode($code: Int!) {
+  mutation JoinQuestByCode($code: Float!) {
     joinQuestByCode(code: $code) {
       code
     }
